backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
Pillow==10.3.0
torch==2.3.0
torchvision==0.18.0
transformers==4.41.2
accelerate==0.30.0
redis==5.0.4
slowapi==0.1.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.1
pydantic-settings==2.2.1
structlog==24.1.0
nltk==3.8.1
numpy==1.26.4
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
